6a/6c/6l: get the imported vanilla plan9 toolchain to compile and build
Wire the vanilla amd64 6a/6c/6l (imported from plan9foundation/plan9.git
in 76125abe7) into goken's build, and fix what's needed to compile:

- convert the three mkfiles from raw plan9-upstream to goken's
  TOP=../../mkconfig/mkfiles/mkone template
- move 6.out.h to include/6.out.h, matching 5.out.h/8.out.h's placement
- reuse existing goken infra instead of vanilla plan9's own: 6a's lex.c
  picks up lexbody/macbody/compat from compilers/cck/ (kencc lineage,
  not Go-era src/cmd/cc/ nor principia's compilers/cc/), 6c links
  against compilers/cck/cc.a (its cc.h already carries the same
  long->int32 and mywait/mycreat->lib9 adaptations this vanilla source
  needs), and 6l reuses linkers/lk/elf.c (32/64-bit, like 7l) instead of
  linkers/8l/elf.c (32-bit only, missing elf64())
- fix latent 64-bit bugs: wput/lput/align/maxround/yylex were typed
  long, correct on native Plan9 where long is 32-bit but wrong on our
  64-bit host (breaks elf.c's Putl and cc.h's prototypes); retyped to
  int32/int, matching the same fix already applied for 7l/8ck
- drop the vanilla fileexists() (l.h decl + compat.c impl): lib9's
  libc.h already provides it, same as 8l
- add assemblers/6a linkers/6l compilers/6c to the top-level mkfile's
  DIRS; a full "mk install" now builds 6a/6c/6l alongside the
  preexisting Go-era 6ag/6cg/6lg

+++ assemblers/6a/a.h
@@ -1,7 +1,7 @@
 #include <u.h>
 #include <libc.h>
 #include <bio.h>
-#include "../6c/6.out.h"
+#include <6.out.h>
 
 
 #ifndef	EXTERN
@@ -139,7 +139,7 @@ void	newfile(char*, int);
 Sym*	slookup(char*);
 Sym*	lookup(void);
 void	syminit(Sym*);
-long	yylex(void);
+int	yylex(void);
 int	getc(void);
 int	getnsc(void);
 void	unget(int);
